Code First Girls - Exercises 3.1 & 3.2
Exercise 3.1
You have a budget of £10 and want to write a program to decide which burger restaurant to go
to.
1. Input the ​ price​ of a burger using ​ input()
2. Check whether the ​ price​ is less than or equal (​ <=​ ) ​ 10.00

Exercise 3.2
Add code to your burger program to input whether the restaurant has a vegetarian option
The output should say whether the cost is within budget ​ AND​ has a vegetarian option. 

Extension:​ Add a check to see it the restaurant's rating is 3 or more.